# chore: Fix consume extension method performance