@neuecc 2025-04-20
Thank you.
There seems to be discussion about JIT optimizing callvirt for constraints, but in reality, it appears to be quite challenging.
The information regarding reference types is interesting.