@neuecc 2025-04-21
I've tuned for ArraySelect.
System.Linq's Array.Select becomes ArraySelectIterator, and one MoveNext is omitted.
Since Select and Where are typical patterns, ZLinq will also add types and introduce optimized paths for arrays.

![image](https://github.com/user-attachments/assets/328f02e0-6e98-43a2-9aa4-5415dd8c6c81)